Changed DepletedCount as a method inside ReadFlockDetails
